#89 update tests to trigger internal errors
